Login-widget has been replaced by reloading page on expired session
